+++ .build/tasks/PlatyPS.UpdateExternalHelp.ps1
@@ -1,4 +1,18 @@
 task updateExternalHelp {
-    Import-Module 'platyPS' -ErrorAction 'Stop'
-    New-ExternalHelp .\Docs\Help -OutputPath .\Source\en-US\ -Force
+    $ProjectName = Get-SamplerProjectName -BuildRoot $BuildRoot
+
+    $OutputDocsDestination = (Get-Item ./output/$ProjectName/**/en-US | Select-Object -Last 1).FullName
+    Import-Module 'Microsoft.PowerShell.PlatyPS' -ErrorAction 'Stop'
+    Measure-PlatyPSMarkdown -Path ./Docs/$ProjectName/*.md |
+        Where-Object Filetype -match 'CommandHelp' |
+        Import-MarkdownCommandHelp -Path {$_.FilePath} |
+        Export-MamlCommandHelp -OutputFolder $OutputDocsDestination -Force
+
+    $ModuleInfo = Import-Module -Name $ProjectName -PassThru -ErrorAction 'Stop'
+    $newMarkdownCommandHelpSplat = @{
+        ModuleInfo = $ModuleInfo
+        OutputFolder = "./Docs"
+        WithModulePage = $false
+    }
+    New-MarkdownCommandHelp @newMarkdownCommandHelpSplat
 }